[Accelerate model loading] Fix meta device and super low memory usage
